refactor path formatter
