+++ lib/fictium/exporters/open_api/v3_exporter.rb
@@ -1,3 +1,5 @@
+require_relative 'v3_exporter/content_formatter'
+require_relative 'v3_exporter/example_formatter'
 require_relative 'v3_exporter/path_formatter'
 require_relative 'v3_exporter/path_generator'
